requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
astor==0.8.1
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
configreader==0.0.6
et-xmlfile==1.1.0
exceptiongroup==1.1.1
execnet==1.9.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
numpy==1.24.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.0
pluggy==1.0.0
py==1.11.0
PySocks==1.7.1
pytest==7.2.2
pytest-ast-transformer==1.0.3
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-soft-assertions==0.1.2
pytest-xdist==3.2.1
python-decouple==3.8
python-dotenv==1.0.0
requests==2.28.2
scipy==1.10.1
selenium==4.8.2
slider==0.7.0
sniffio==1.3.0
sortedcontainers==2.4.0
texttable==1.6.7
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
webdriver-manager==3.8.5
workbook==1.1
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
